+++ packages/aft/bin/aft.dart
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import 'dart:io';
+
 import 'package:aft/aft.dart';
 import 'package:args/command_runner.dart';
 
@@ -23,8 +25,35 @@ Future<void> main(List<String> args) async {
       help: 'Prints verbose logs',
       defaultsTo: false,
     )
+    ..argParser.addOption(
+      'directory',
+      help: 'Directory to run commands from. Defaults to current directory.',
+      hide: true,
+    )
     ..addCommand(GenerateSdkCommand())
     ..addCommand(ListPackagesCommand())
-    ..addCommand(DepsCommand());
-  await runner.run(args);
+    ..addCommand(DepsCommand())
+    ..addCommand(LinkCommand())
+    ..addCommand(CleanCommand())
+    ..addCommand(PubCommand())
+    ..addCommand(BootstrapCommand());
+  try {
+    await runner.run(args);
+  } on UsageException catch (e) {
+    stderr
+      ..writeln(e.message)
+      ..writeln(e.usage);
+    exitCode = 1;
+  } on Object catch (e, st) {
+    stderr
+      ..writeln(e.toString())
+      ..writeln(st);
+    exitCode = 1;
+  } finally {
+    // Free up resources before exiting..
+    for (final command in runner.commands.values.whereType<AmplifyCommand>()) {
+      command.close();
+    }
+    exit(exitCode);
+  }
 }