+++ packages/worker_bee/e2e_test/lib/e2e_test.dart
@@ -1,0 +1,2 @@
+/// E2E tests for the `worker_bee` package.
+library e2e_test;