+++ packages/smithy/smithy_aws/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.1
+
+- Fix package analysis issues
+
 ## 0.1.0
 
 - Initial version.